fix: api 수정
