+++ web/app/onboarding/page.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { useRouter } from 'next/navigation';
 import OnboardingCarousel from '@/components/onBoarding/OnboardingCarousel';
-
+import { useSearchParams } from 'next/navigation';
 function Page() {
   const slides: { title: string; text: string; image: string }[] = [
     {
